Spatial Filters/requirements.txt
numpy
matplotlib
pillow==10.2.0
opencv-python==4.9.0.80